feat(build): add emscripten support

+++ src/Stream.cc
@@ -73,7 +73,7 @@ ZkSize ZkRead_getSize(ZkRead* slf) {
 	auto off = slf->tell();
 	slf->seek(0, zenkit::Whence::END);
 	auto size = slf->tell();
-	slf->seek(static_cast<int64_t>(off), zenkit::Whence::BEG);
+	slf->seek(static_cast<ssize_t>(off), zenkit::Whence::BEG);
 	return size;
 }
 
@@ -82,6 +82,6 @@ ZkSize ZkRead_getBytes(ZkRead* slf, void* buf, ZkSize length) {
 	auto off = slf->tell();
 	slf->seek(0, zenkit::Whence::BEG);
 	auto count = slf->read(buf, length);
-	slf->seek(static_cast<int64_t>(off), zenkit::Whence::BEG);
+	slf->seek(static_cast<ssize_t>(off), zenkit::Whence::BEG);
 	return count;
 }
